requirements.txt
discord.py
asyncio
pafy
ydl
pip
youtube-dl
ffmpeg-python